clean up unused functions, make stringify python package
