+++ pkg/workloads/cortex/lib/test/util_test.py
@@ -64,46 +64,3 @@ def test_is_number_col():
     assert util.is_number_col([None, 1, None])
     assert util.is_number_col([None, 1.1, None])
     assert not util.is_number_col([None, None, None])
-
-
-def test_print_samples_horiz(caplog):
-    caplog.set_level(logging.INFO)
-
-    samples = [
-        {
-            "test1": float(120),
-            "test2": 0.2,
-            "testlongstring": 0.20,
-            "a": 0.23,
-            "b": 0.233333333333333333,
-        },
-        {
-            "test1": 11,
-            "test2": 18,
-            "testreallyreallyreallyreallyreallylongstring": 18,
-            "a": -12,
-            "b": 10,
-        },
-        {
-            "test1": 13,
-            "test2": 13,
-            "testlongstring": 13,
-            "testreallyreallyreallyreallyreallylongstring": 13,
-            "a": 133333,
-            "b": None,
-        },
-    ]
-    util.print_samples_horiz(samples)
-
-    records = [r.message for r in caplog.records]
-
-    expected = (
-        "a:                      0.23, -12, 133333\n"
-        "b:                      0.23,  10,\n"
-        "test1:                120.00,  11,     13\n"
-        "test2:                  0.20,  18,     13\n"
-        "testlongstring:         0.20,    ,     13\n"
-        "testreallyreallyr...:       ,  18,     13\n"
-    )
-
-    assert "\n".join(records) + "\n" == expected
