+++ pkg/workloads/cortex/onnx_serve/api.py
@@ -27,6 +27,7 @@ from cortex.lib import util, package, Context, api_utils
 from cortex.lib.storage import S3
 from cortex.lib.log import get_logger, print_obj
 from cortex.lib.exceptions import CortexException, UserRuntimeException, UserException
+from cortex.lib.stringify import json_tricks_encoder, print_obj, to_string
 
 
 def cortex_print(*args, **kwargs):
@@ -40,7 +41,7 @@ logger.propagate = False  # prevent double logging (flask modifies root logger)
 
 app = Flask(__name__)
 
-app.json_encoder = util.json_tricks_encoder
+app.json_encoder = json_tricks_encoder
 
 # https://github.com/microsoft/onnxruntime/blob/v0.4.0/onnxruntime/python/onnxruntime_pybind_mlvalue.cc
 ONNX_TO_NP_TYPE = {
@@ -175,7 +176,7 @@ def predict(app_name, api_name):
     if not util.is_dict(payload) or "samples" not in payload:
         message = 'top level "samples" key not found in request'
         if debug:
-            message += "; payload: {}".format(util.pp_str_flat(payload))
+            message += "; payload: {}".format(to_string(payload, flat=True))
         return prediction_failed(message)
 
     predictions = []
@@ -284,13 +285,13 @@ def start(args):
         local_cache["input_metadata"] = sess.get_inputs()
         logger.info(
             "input_metadata: {}".format(
-                util.pp_str_flat(extract_signature(local_cache["input_metadata"]))
+                to_string(extract_signature(local_cache["input_metadata"]), flat=True)
             )
         )
         local_cache["output_metadata"] = sess.get_outputs()
         logger.info(
             "output_metadata: {}".format(
-                util.pp_str_flat(extract_signature(local_cache["output_metadata"]))
+                to_string(extract_signature(local_cache["output_metadata"]), flat=True)
             )
         )
     except CortexException as e:
